--- sml/fyy_pca/BUILD.bazel
@@ -1,17 +1,0 @@
-load("//bazel:spu.bzl", "sml_py_library")
-
-package(default_visibility = ["//visibility:public"])
-
-sml_py_library(
-    name = "jacobi_evd",
-    srcs = ["jacobi_evd.py"],
-)
-
-sml_py_library(
-    name = "jacobi_pca",
-    srcs = ["jacobi_pca.py"],
-    deps = [
-        ":jacobi_evd",
-        "//sml/utils:extmath",
-    ],
-)
