--- sml/fyy_pca/emulations/3pc_128.json
@@ -1,52 +1,0 @@
-{
-    "id": "outsourcing.3pc",
-    "nodes": {
-        "node:0": "127.0.0.1:61920",
-        "node:1": "127.0.0.1:61921",
-        "node:2": "127.0.0.1:61922",
-        "node:3": "127.0.0.1:61923",
-        "node:4": "127.0.0.1:61924"
-    },
-    "devices": {
-        "SPU": {
-            "kind": "SPU",
-            "config": {
-                "node_ids": [
-                    "node:0",
-                    "node:1",
-                    "node:2"
-                ],
-                "spu_internal_addrs": [
-                    "127.0.0.1:61930",
-                    "127.0.0.1:61931",
-                    "127.0.0.1:61932"
-                ],
-                "experimental_data_folder": [
-                    "/tmp/spu_data_0/",
-                    "/tmp/spu_data_1/",
-                    "/tmp/spu_data_2/"
-                ],
-                "runtime_config": {
-                    "protocol": "ABY3",
-                    "field": "FM128",
-                    "fxp_fraction_bits": 30,
-                    "enable_pphlo_profile": false,
-                    "enable_hal_profile": false,
-                    "enable_pphlo_trace": false
-                }
-            }
-        },
-        "P1": {
-            "kind": "PYU",
-            "config": {
-                "node_id": "node:3"
-            }
-        },
-        "P2": {
-            "kind": "PYU",
-            "config": {
-                "node_id": "node:4"
-            }
-        }
-    }
-}
