--- sml/fyy_pca/emulations/BUILD.bazel
@@ -1,27 +1,0 @@
-load("//bazel:spu.bzl", "sml_py_binary")
-
-package(default_visibility = ["//visibility:public"])
-
-sml_py_binary(
-    name = "jacobipca_emul",
-    srcs = ["jacobipca_emul.py"],
-    deps = [
-        "//sml/fyy_pca:jacobi_pca",
-    ],
-)
-
-sml_py_binary(
-    name = "jacobievd_emul",
-    srcs = ["jacobievd_emul.py"],
-    data = [":conf"],
-    deps = [
-        "//sml/fyy_pca:jacobi_evd",
-    ],
-)
-
-filegroup(
-    name = "conf",
-    srcs = [
-        "3pc_128.json",
-    ],
-)
