--- sml/fyy_pca/tests/BUILD.bazel
@@ -1,19 +1,0 @@
-load("//bazel:spu.bzl", "sml_py_test")
-
-package(default_visibility = ["//visibility:public"])
-
-sml_py_test(
-    name = "jacobipca_test",
-    srcs = ["jacobipca_test.py"],
-    deps = [
-        "//sml/fyy_pca:jacobi_pca",
-    ],
-)
-
-sml_py_test(
-    name = "jacobievd_test",
-    srcs = ["jacobievd_test.py"],
-    deps = [
-        "//sml/fyy_pca:jacobi_evd",
-    ],
-)
