+++ sml/utils/tests/BUILD.bazel
@@ -22,3 +22,11 @@ sml_py_test(
         "//sml/utils:extmath",
     ],
 )
+
+sml_py_test(
+    name = "jacobievd_test",
+    srcs = ["jacobievd_test.py"],
+    deps = [
+        "//sml/utils:extmath",
+    ],
+)
